% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_family_tau.R
\name{accept_family_tau}
\alias{accept_family_tau}
\title{Curate Quantile Specific Acceptable Families}
\usage{
accept_family_tau(post_Q_tau, XX, indicators, eps_level = 0.05, eta_level = 0)
}
\arguments{
\item{post_Q_tau}{posterior samples of the conditional quantiles at \code{XX}}

\item{XX}{covariate values}

\item{indicators}{inclusion indicators from BBA algorithn}

\item{eps_level}{slack variable}

\item{eta_level}{slack variable}
}
\value{
\eqn{\tau}-specific acceptable families
}
\description{
Curate Quantile Specific Acceptable Families
}
