% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_and_bound.R
\name{branch_and_bound}
\alias{branch_and_bound}
\title{Branch-and-bound algorithm for linear subset search}
\usage{
branch_and_bound(
  yy,
  XX,
  wts = NULL,
  n_best = 15,
  to_include = 1,
  to_exclude = NULL,
  searchtype = "exhaustive"
)
}
\arguments{
\item{XX}{matrix of covariates}

\item{wts}{vector of observation weights (for weighted least squares)}

\item{n_best}{number of "best" subsets for each model size}

\item{to_include}{indices of covariates to include in \emph{all} subsets}

\item{to_exclude}{indices of covariates to exclude from \emph{all} subsets}

\item{searchtype}{use exhaustive search, forward selection, backward selection or sequential replacement to search}

\item{Q_hat_tau}{posterior mean of the \eqn{\tau}-conditional quantile function at \code{XX}}
}
\value{
\code{inclusion_index}: the matrix of inclusion indicators (columns) for
each subset returned (rows)
}
\description{
Search for the "best" (according to residual sum of squares)
linear subsets of each size. The algorithm may collect
the \code{n_best} "best" subsets of each size, include or
exclude certain variables automatically, and apply
forward, backward, or exhaustive search.
}
