% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loss.R
\name{check_loss}
\alias{check_loss}
\title{Compute the check loss between response \code{y} and quantile-predictions \code{qhat} for any quantile \code{tau}}
\usage{
check_loss(y, qhat, tau)
}
\arguments{
\item{y}{response values}

\item{qhat}{quantile predictions}

\item{tau}{quantile of interest}
}
\value{
the average check loss across all quantile predictions
}
\description{
Compute the check loss between response \code{y} and quantile-predictions \code{qhat} for any quantile \code{tau}
}
