% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_loss_l2.R
\name{post_loss_l2}
\alias{post_loss_l2}
\title{Compute the posterior aggregated squared-error loss}
\usage{
post_loss_l2(post_Q_tau, XX, indicators)
}
\arguments{
\item{post_Q_tau}{\code{S x n} matrix of S posterior samples of the conditional quantile function
\eqn{Q_{\tau}(x_i)} for each covariate value in \code{XX}}

\item{XX}{\code{n x p} matrix of covariates at which to evaluate subsets}

\item{indicators}{\code{L x p} matrix of inclusion indicators extracted from the branch and bound
search algorithm. Each row corresponds to a candidate subset, with the indicator determining whether \code{X_j} is included.}
}
\value{
a list with four elements \code{post_loss}, \code{post_loss_raw},
\code{ref}, and \code{beta_hat_L}
}
\description{
Use posterior draws of the conditional quantile function to compute the residual sum
of squares (RSS) between any subset of predictors and the model-fitted quantiles. These quantities
are computed given filtering of subsets from the branch and bound search algorithm, and the
and posterior samples of conditional quantile function from the overarching Bayesian model fit to
the data.
}
\details{
\code{post_loss} is an \code{S x L} matrix including posterior samples of the difference in RSS between the anchor,
which is the posterior mean of the conditional quantile function under the unifying Bayesian model, and each subset from
the branch and bound filtering. This is used to determine the acceptable family.
\code{post_loss_raw} is an \code{S x L} matrix with the posterior samples of the RSS for each subset in the branch and bound filtering.
\code{ref} is an \code{S x 1} array containing posterior samples of the RSS for the anchor action. \code{post_loss_raw} and \code{ref} are
the constituent terms to compute \code{post_loss}. Finally,
\code{beta_hat_L} is the optimal action for each subset in the branch and bound filtering
}
