% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_posterior.R
\name{proj_posterior}
\alias{proj_posterior}
\title{Projected posterior distribution for regression coefficients, i.e. the posterior action}
\usage{
proj_posterior(post_Q_tau, XX, sub_x = 1:ncol(XX))
}
\arguments{
\item{post_Q_tau}{\code{S x n} matrix of posterior samples of the conditional quantile function at each design point among \code{XX} covariate values}

\item{XX}{\code{n x p} matrix of covariates}

\item{sub_x}{vector of inclusion indicators for the \code{p} covariates;
the remaining coefficients will be fixed at zero}
}
\value{
\code{post_beta}: the \code{S x p} matrix of
draws from the projected posterior distribution of the regression coefficients.
}
\description{
Given draws from the posterior distribution of the conditional quantile function, project these
draws onto (a subset of) the covariates. This produces
many posterior draws for the regression coefficients,
which provides uncertainty quantification.
}
