% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\title{Variable importance for each quantile-specific acceptable family}
\usage{
var_imp(indicators, all_accept, co = TRUE, xnames = NULL)
}
\arguments{
\item{indicators}{\code{L x p} matrix of inclusion indicators (booleans)
where each row denotes a candidate subset (from \code{branch_and_bound})}

\item{all_accept}{indices (i.e., rows of \code{indicators})
that correspond to the acceptable subsets (from \code{accept_family_tau})}

\item{co}{logical; if TRUE, compute and return the co-variable importances}

\item{xnames}{the names of the x-variables}
}
\value{
a list with the variable importances \code{vi_inc} and the co-variable
importances \code{vi_co}
}
\description{
Given the candidate subsets and the indicators of any quantile-specific acceptable family
of subsets, compute for each variable the proportion of acceptable subsets in
which that variable appears. If specified, variable co-appearances
can be computed as reported as well.
}
