% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_family_tau.R
\name{accept_family_tau}
\alias{accept_family_tau}
\title{Curate Quantile Specific Acceptable Families}
\usage{
accept_family_tau(post_Q_tau, XX, indicators, eps_level = 0.05, eta_level = 0)
}
\arguments{
\item{post_Q_tau}{\code{S x n} matrix of S posterior samples of the conditional quantile function
\eqn{Q_{\tau}(x_i)} for each covariate value in \code{XX}}

\item{XX}{\code{n x p} matrix of covariates at which to evaluate subsets}

\item{indicators}{\code{L x p} matrix of inclusion indicators extracted from the branch and bound}

\item{eps_level}{slack variable \eqn{\epsilon}, used to determine member subsets in the acceptable family. Decrease to discover more acceptable subsets.}

\item{eta_level}{slack variable \eqn{\eta}, used to determine member subsets in the acceptable family. Decrease to discover more acceptable subsets.}
}
\value{
A list containing \code{all_accept} which are the indices of the subsets from \code{indicators} which are acceptable,
\code{ell_small}, which the index of the smallest acceptable subset, and \code{beta_hat_small} which is the optimal action (\eqn{\tau} quantile regression coefficients) for the smallest acceptable subset.
}
\description{
Using posterior samples of the \eqn{\tau}th conditional quantile function, evaluate the posterior probability that each candidate
subset obtained from the branch and bound search algorithm has comparable predictive power to the anchoring action. The anchor is
chosen as the ``best" fit to the model-based conditional quantile function, which we set as the model-based posterior mean.
}
\details{
For any subset \eqn{S} obtained by the quantile-specific branch and bound search,
that subset is deemed acceptable if \eqn{P(D^{\tau}_{S, \hat{Q}}  \leq \eta) \geq \epsilon}, where \eqn{D^{\tau}_{S, \hat{Q}} = 100 \ \code{x} \ (L^{\tau}_{S}(\theta)  -L^{\tau}_{\hat{Q}}(\boldsymbol\theta) )/L^{\tau}_{\hat{Q}}(\theta) },
\eqn{\hat Q} is the posterior mean of \eqn{\tau}th conditional quantile function, \eqn{L} is the aggregated \eqn{L^2} loss, and \eqn{\theta} are the Bayesian regression model parameters, drawn from the posterior. The smallest acceptable subset
is then the smallest in cardanlity subset collected that maintains satisfactory predictive power under the criteria outlined above.
}
